+++ src/main/java/ru/yandex/practicum/filmorate/storage/db/DirectorMapper.java
@@ -1,0 +1,18 @@
+package ru.yandex.practicum.filmorate.storage.db;
+
+import org.springframework.jdbc.core.RowMapper;
+import ru.yandex.practicum.filmorate.model.Director;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+// service class
+class DirectorMapper implements RowMapper<Director> {
+    @Override
+    public Director mapRow(ResultSet rs, int rowNum) throws SQLException {
+        return Director.builder()
+                .id(rs.getInt("director_id"))
+                .name(rs.getString("name"))
+                .build();
+    }
+}
